rename contract to dto
