+++ src/loan-service/src/main/java/com/asc/loanservice/domain/loan/application/LoanEvaluatorProviderPort.java
@@ -1,6 +1,6 @@
 package com.asc.loanservice.domain.loan.application;
 
-import com.asc.loanservice.domain.loan.application.contract.LoanApplicationRequest;
+import com.asc.loanservice.domain.loan.application.dto.LoanApplicationRequest;
 
 
 public interface LoanEvaluatorProviderPort {
