+++ src/loan-service/src/main/java/com/asc/loanservice/domain/loan/application/LoanMapper.java
@@ -1,7 +1,7 @@
 package com.asc.loanservice.domain.loan.application;
 
-import com.asc.loanservice.domain.loan.application.contract.LoanApplicationCreateInput;
-import com.asc.loanservice.domain.loan.application.contract.LoanApplicationRequest;
+import com.asc.loanservice.domain.loan.application.dto.LoanApplicationCreateInput;
+import com.asc.loanservice.domain.loan.application.dto.LoanApplicationRequest;
 import org.mapstruct.Mapper;
 import org.springframework.stereotype.Component;
