+++ src/loan-service/src/main/java/com/asc/loanservice/domain/loan/evaluation/Rule.java
@@ -1,6 +1,6 @@
 package com.asc.loanservice.domain.loan.evaluation;
 
-import com.asc.loanservice.domain.loan.application.contract.LoanApplicationRequest;
+import com.asc.loanservice.domain.loan.application.dto.LoanApplicationRequest;
 
 
 public interface Rule {
